wave 3 part 1

+++ app/controllers/tasks_controller.rb
@@ -30,11 +30,15 @@ class TasksController < ApplicationController
 
         redirect_to tasks_path
     end
-    
+
         #from wave 4 test #1- done in class 
     def destroy
         task_id = params[:id]
         task = Task.find_by(id: task_id)
+        unless task
+            head :not_found
+            return
+        end
         task.destroy
         redirect_to tasks_path
     end
